Reorganizo bastantes cosas
