+++ scripts/release.js
@@ -1,6 +1,6 @@
 import { $ } from 'bun';
-import { dirname } from 'path';
 import { parseArgs } from 'util';
+import { copyTrackedFiles, compressAndGenerateChecksums, exportMakefile } from './utils';
 
 const { values } = parseArgs({
   args: Bun.argv,
@@ -17,15 +17,11 @@ const { values } = parseArgs({
       type: 'string',
       default: 'dist'
     },
-    templates: {
-      type: 'string',
-      default: 'project,static,shared'
-    },
   },
   allowPositionals: true,
 });
 
-const main = async ({ tag, src, dest, templates }) => {
+const main = async ({ tag, src, dest }) => {
   console.log(`cleaning up ${dest}...\n\n`);
 
   await $`rm -rfv ${dest}`;
@@ -33,38 +29,26 @@ const main = async ({ tag, src, dest, templates }) => {
 
   console.log(`\n\nparsing templates from ${src}...`);
 
-  for (const template of templates.split(',')) {
+  for (const template of ['project', 'static', 'shared']) {
     console.log(`\n\nparsing ${template}...\n\n`);
 
-    await $`rsync -r --exclude-from=${src}/${template}/.gitignore ${src}/${template} ${dest}`;
-    await $`rsync -r ${src}/${template}/.vscode ${dest}/${template}`
+    await copyTrackedFiles(`${src}/templates/${template}`, dest);
 
-    const makefile = await parseMakefile(`${src}/${template}/makefile`);
-    await $`echo ${makefile} | tee makefile`.cwd(`${dest}/${template}`);
+    await $`cp -rv ${src}/configs/vscode ${dest}/${template}/.vscode`;
+    if (template !== 'project') {
+      await $`rm -fv ${dest}/${template}/.vscode/launch.json`;
+    }
+
+    await exportMakefile(`${src}/templates/${template}`, `${dest}/${template}`);
 
     console.log(`packing ${template} with tag ${tag}...\n\n`);
 
-    await $`tar -czvf ${template}-${tag}.tar.gz ${template}`.cwd(dest)
-    await $`md5sum ${template}-${tag}.tar.gz > ${template}-${tag}.tar.gz.md5`.cwd(dest);
-    await $`sha1sum ${template}-${tag}.tar.gz > ${template}-${tag}.tar.gz.sha1`.cwd(dest);
+    await compressAndGenerateChecksums(tag, `${dest}/${template}`);
 
     console.log(`\n\ncleaning up ${template}...\n\n`);
 
-    await $`rm -rfv ${template}`.cwd(dest)
+    await $`rm -rfv ${template}`.cwd(dest);
   }
 };
-
-const parseMakefile = async (file) => {
-  const lines = []
-  for await(const line of $`cat ${file}`.lines()) {
-    lines.push(await parseMakefileLine(file, line));
-  }
-  return lines.join('\n');
-}
-
-const parseMakefileLine = async (file, line) => {
-  const [_, include] = line.split('include ../');
-  return include ? $`cat ${dirname(file)}/../${include}`.text() : line;
-}
 
 await main(values);
