+++ src/configs/vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "debug.onTaskErrors": "abort",
+  "files.associations": {
+    "*.h": "c"
+  },
+  "C_Cpp.errorSquiggles": "disabled"
+}
