--- src/project/.vscode/launch.json
@@ -1,51 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?linkid=830387
-  // for the documentation about the launch.json format
-  "version": "0.2.0",
-  "configurations": [
-    // #region run
-    {
-      "name": "run",
-      "type": "cppdbg",
-      "request": "launch",
-      "program": "${workspaceFolder}/bin/${workspaceFolderBasename}",
-      "args": [
-        // TODO: Agregar los argumentos que se necesiten
-      ],
-      "stopAtEntry": false,
-      "cwd": "${workspaceFolder}",
-      "environment": [],
-      "externalConsole": false,
-      "MIMode": "gdb",
-      "setupCommands": [
-        {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
-        }
-      ],
-      "preLaunchTask": "build"
-    },
-    // #endregion run
-    {
-      "name": "test",
-      "type": "cppdbg",
-      "request": "launch",
-      "program": "${workspaceFolder}/bin/${workspaceFolderBasename}_tests",
-      "args": [],
-      "stopAtEntry": false,
-      "cwd": "${workspaceFolder}",
-      "environment": [],
-      "externalConsole": false,
-      "MIMode": "gdb",
-      "setupCommands": [
-        {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
-        }
-      ],
-      "preLaunchTask": "build"
-    }
-  ]
-}
