--- src/project/.vscode/tasks.json
@@ -1,28 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "2.0.0",
-  "tasks": [
-    // #region build
-    {
-      "label": "build",
-      "command": "make all",
-      "type": "shell",
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      },
-      "problemMatcher": ["$gcc"]
-    },
-    // #endregion build
-    {
-      "label": "clean",
-      "command": "make clean",
-      "type": "shell",
-      "group": {
-        "kind": "none"
-      },
-      "problemMatcher": []
-    }
-  ]
-}
