--- src/shared/.vscode/c_cpp_properties.json
@@ -1,19 +1,0 @@
-{
-  // See https://code.visualstudio.com/docs/cpp/c-cpp-properties-schema-reference
-  // for the documentation about the c_cpp_properties.json format
-  "configurations": [
-    {
-      "name": "Linux",
-      "includePath": [
-        "${workspaceFolder}/src"
-        // TODO: Agregar los directorios de las bibliotecas a utilizar
-      ],
-      "defines": [],
-      "compilerPath": "/usr/bin/gcc",
-      "cStandard": "gnu17",
-      "cppStandard": "gnu++17",
-      "intelliSenseMode": "linux-gcc-x64"
-    }
-  ],
-  "version": 4
-}
