--- src/static/.vscode/tasks.json
@@ -1,26 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "label": "build",
-      "command": "make all",
-      "type": "shell",
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      },
-      "problemMatcher": ["$gcc"]
-    },
-    {
-      "label": "clean",
-      "command": "make clean",
-      "type": "shell",
-      "group": {
-        "kind": "none"
-      },
-      "problemMatcher": []
-    }
-  ]
-}
