+++ src/templates/project/.gitignore
@@ -17,4 +17,5 @@ RemoteSystemsTempFiles/
 .vscode/*
 !.vscode/c_cpp_properties.json
 !.vscode/launch.json
+!.vscode/settings.json
 !.vscode/tasks.json
