+++ src/templates/project/settings.mk
