+++ src/templates/project/src/main.c
