+++ src/templates/shared/settings.mk
