+++ src/templates/shared/src/shared/hello.c
