+++ src/templates/static/makefile
