+++ src/templates/static/settings.mk
