+++ src/templates/static/src/utils/hello.c
